requirements.txt
esp32-machine-emulator==1.1.3
esptool==3.1
adafruit-ampy==1.1.0
black==21.9b0
paho-mqtt==1.6.1
isort==5.9.3
